+++ railties/test/generators/app_generator_test.rb
@@ -113,17 +113,17 @@ class AppGeneratorTest < Rails::Generators::TestCase
 
   def test_invalid_javascript_option_raises_an_error
     content = capture(:stderr) { run_generator([destination_root, "-j", "unknown"]) }
-    assert_match(/Invalid value for --javascript option/, content)
+    assert_match(/Expected '--javascript' to be one of/, content)
   end
 
   def test_invalid_asset_pipeline_option_raises_an_error
     content = capture(:stderr) { run_generator([destination_root, "-a", "unknown"]) }
-    assert_match(/Invalid value for --asset-pipeline option/, content)
+    assert_match(/Expected '--asset-pipeline' to be one of/, content)
   end
 
   def test_invalid_css_option_raises_an_error
     content = capture(:stderr) { run_generator([destination_root, "-c", "unknown"]) }
-    assert_match(/Invalid value for --css option/, content)
+    assert_match(/Expected '--css' to be one of/, content)
   end
 
   def test_application_job_file_present