+++ railties/CHANGELOG.md
@@ -1,3 +1,11 @@
+*   Disallow invalid values for rails new options.
+
+    The `--database`, `--asset-pipeline`, `--css`, and `--javascript` flags for
+    `rails new` can all take different options. This change adds checks to
+    options to make sure the user enters the correct value.
+
+    *Tony Drake*, *Akhil G Krishnan*, *Petrik de Heus*
+
 *   Conditionally print `$stdout` when invoking `run_generator`
 
     In an effort to improve the developer experience when debugging